Crate genfsm

Structs

StateMachine
The core of the GenFSM library. Holds the Executor and current state of a state machine, and handles the processing of NextState values from Executors.

Enums

NextState
Represents the next state an FSM will enter

Type Definitions

Executor
An input handler that executes changes on a state