Enum genfsm::NextState

pub enum NextState<S, I> {
    ChangeState(Executor<S, I>, S),
    Continue(S),
    End(S),
}

Represents the next state an FSM will enter

Variants

ChangeState

Change the state and the executor that handles input

Continue

Continue with the same executor

End

FSM has reached completion. Any further attempts to handle input will result in no changes